requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distro==1.9.0
idna==3.10
lxml==5.3.2
numpy==2.2.4
pandas==2.2.3
pdfminer.six==20250327
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tabula-py==2.10.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
